# internal/dto/short_url_report_dto.go
package dto

import "time"

// ShortURLReportDto represents the short URL for this application with visiting statistics
//
// swagger:model ShortURLReportDto
type ShortURLReportDto struct {
	// min: 1
	ID int64 `json:"id"`

	// unique: true
	LongURL string `json:"long_url" db:"long_url"`

	// unique: true
	Token string `json:"token"`

	Enabled bool `json:"enabled"`

	CreatedAt time.Time `json:"created_at" db:"created_at"`

	Visits int64 `json:"visits"`
}